server info: make it versioned and allow null command in omcache_stat
Server info struct now has a version number to make it possible to add
members to it in a backwards compatible way.  It also gained a
'server_index' member which is the OMcahce internal server id and can be
passed on to omcache_stat and similar functions.

Passing NULL to omcache_stat is now equivalent to passing it an empty string
which, on most memcached server implementations, returns the default set of
statistics.