request lookup: greatly simplified
Modified omcache_io to not modify request array at all and to check that
it's ordered properly, it can now only contain requests in req_id order with
no gaps which is what omcache_command() sets up.  Any other users of it will
have to make sure they pass in a properly ordered request array.

This simplifies request lookup logic a lot as we don't have to do any
lookups in the request list to see if incoming responses are interesting,
just req_id_min and req_id_max are enough for that.