buffering: fixes for handling large and too large responses
Only require the missing part of the message to be read instead of requiring
the buffer to be large enough to hold the existing parts plus the full
message.

If the buffer is still too small to hold the full value store a 'failure
value' in the response lookup object when needed and make sure we don't set
key in the failure value if the buffer is too small to hold it.