server timeouts: fix dead timeout comparison
The dead timeout (msec) is the time at which, not after which, a server will
be marked dead.  This is important because we calculate poll timeouts using
the same setting and if we didn't use >= comparison tests would sometimes
fail as it takes twice as long as it should for a server to be marked dead.

Also, add a 0.1 sleep to memcached suspend test to allow SIGSTOP to be
delivered to the target process.