omcache: fix coverity findings
Fix a potential double freeaddrinfo() call in a rare case where
getaddrinfo() fails after we've connected to a server and then failed to
send a noop there.

Also Silence a couple of false alarms by using a 1-element list for requests
instead of passing the address of the request struct around and make sure to
zero out value in omcache_get() to not return invalid flags or cas.