commands: add append and prepend commands
Added a new error code OMCACHE_NOT_STORED which maps to memcached's
PROTOCOL_BINARY_RESPONSE_NOT_STORED which is issued for appends and prepends
to non-existent keys.  Also added omcache.py mapping for TooLargeValueError
and DeltaBadValueError which were introduced earlier on but lacked mappings
in the Python bindings.

As version 0.1.0 was already tagged add the new symbols to OMCACHE_0.2 which
will be the next release.