python delta operations: don't initialize value by default
Delta operation definition in the protocol is a bit weird; if 'expiration'
is set to DELTA_NO_ADD (0xffffffff) the value will not be created if it
doesn't exist yet, but since we have a 'initial' argument in the python api
we'd really like to use it to signal whether or not a value should be
initialized.

To do that we'll map expiration to DELTA_NO_ADD if initial is None and
expiration is empty and throw an error if initial is None but expiration
isn't empty.