commands: return OMCACHE_BUFFERED if we haven't received a response
Previously commands would return OMCACHE_OK if timeout_msec == 0 and data
was successfully written to the backends.  This was misleading, as there's
no guarantees that the backends have processed (or even received!) the
request at that point.  Return OMCACHE_BUFFERED instead to tell the client
that they need to call omcache_io() before they can be sure the data has hit
the backends.