omcache: keep active requests in a hash table and make sure they're handled
Replaced the previous naive request / response handling implementation of
omcache_command() + omcache_io() where we're just passing around a struct
with minimum and maximum request id and hoping everything goes well with an
actual hash table of active requests.  Whenever we find a response matching
an active request or decide to offline a server all requests that were
pending will be marked as handled.

This allows omcache_io() to report proper status back to the caller.
Previously omcache_io() would keep polling if we were trying to connect to a
disabled server even if all requests were already answered, it would also
return OMCACHE_OK always after there were no more sockets to poll.  This was
incorrect, as we may end up with no sockets to poll because we had to
offline some backends because they failed asynchronously, for example in
connect or name resolution phases.