omcache.py: refactored code to enable gevent support
Don't use the functions from commands.c which either block or require using
a response callback, but instead create the request structures in the python
code and pass them to omcache_command() and omcache_io() to allow custom
select implementation to wait on the sockets.

This allows OMcache to be used with gevent by passing gevent.select.select
to OMcache() constructor as 'select'.