python: added check-pylint makefile target and fixed the issues
Run pylint on Python code and fix the issues it found and silence the false
positives that came up.  Specifically we need to disable warnings on unused
variables when calling OMcache._request() which allocates C objects which
are not otherwise tracked in the Python side and would end up being garbage
collected before we use them.