tests: more verbose logging and ability to select case in args
Set Check's output verbosity to VERBOSE up from NORMAL (prints a summary of
all test cases at the end of the run) and make it log test execution on
stdout.  Also make sure stderr and stdout are linebuffered to make sure the
output is at least somewhat ordered.

Also made test_omcache interpret the first argument (if any) as the name of
the test case to run.  The default is to run all cases.